Refactor LogixNG to move "Browse" to jmri.jmrit.logixng.tools.swing
