+++ java/src/jmri/jmrit/beantable/AbstractLogixNGTableAction.java
@@ -2,7 +2,6 @@ package jmri.jmrit.beantable;
 
 import java.awt.*;
 import java.awt.event.*;
-import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
@@ -22,11 +21,9 @@ import jmri.UserPreferencesManager;
 import jmri.jmrit.logixng.Base;
 import jmri.jmrit.logixng.tools.swing.AbstractLogixNGEditor;
 import jmri.jmrit.logixng.tools.swing.DeleteBean;
-import jmri.util.FileUtil;
+import jmri.jmrit.logixng.tools.swing.LogixNGBrowseWindow;
 import jmri.util.JmriJFrame;
 import jmri.util.swing.JmriJOptionPane;
-
-import org.apache.commons.lang3.mutable.MutableInt;
 
 /**
  * Swing action to create and register a LogixNG Table.
@@ -51,15 +48,6 @@ public abstract class AbstractLogixNGTableAction<E extends NamedBean> extends Ab
     private static final ResourceBundle rbx = ResourceBundle.getBundle("jmri.jmrit.logixng.LogixNGBundle");
     private static final ResourceBundle rbx2 = ResourceBundle.getBundle("jmri.jmrit.logixng.tools.swing.LogixNGSwingBundle");
 
-    // Browser Options
-    static final String PRINT_LINE_NUMBERS_OPTION = "jmri.jmrit.logixng.PrintLineNumbers";
-    static final String PRINT_ERROR_HANDLING_OPTION = "jmri.jmrit.logixng.ErrorHandling";
-    static final String PRINT_NOT_CONNECTED_OPTION = "jmri.jmrit.logixng.NotConnectedSockets";
-    static final String PRINT_LOCAL_VARIABLES_OPTION = "jmri.jmrit.logixng.LocalVariables";
-    static final String PRINT_SYSTEM_NAMES_OPTION = "jmri.jmrit.logixng.SystemNames";
-    static final String PRINT_DISABLED_OPTION = "jmri.jmrit.logixng.Disabled";
-
-    JTextArea _textContent;
     DeleteBean<E> deleteBean = new DeleteBean<>(getManager());
 
     /**
@@ -97,7 +85,7 @@ public abstract class AbstractLogixNGTableAction<E extends NamedBean> extends Ab
 
     protected boolean browseMonoSpace() { return false; }
 
-    protected abstract String getBeanText(E bean);
+    protected abstract String getBeanText(E bean, Base.PrintTreeSettings printTreeSettings);
 
     protected abstract String getBrowserTitle();
 
@@ -140,18 +128,9 @@ public abstract class AbstractLogixNGTableAction<E extends NamedBean> extends Ab
      */
     @Override
     public void setMenuBar(BeanTableFrame<E> f) {
-        JMenuBar menuBar = f.getJMenuBar();
-
-        JMenu menu = menuBar.getMenu(0);   // Menu "File"
-        JMenuItem r2 = new JMenuItem(Bundle.getMessage("BrowseAllLogixNGs"));  // NOI18N
-        r2.addActionListener((ActionEvent e) -> {
-            getPrintTreeSettings();
-            makeBrowserWindow(true);
-        });
-        menu.add(r2);
-
-        menu = new JMenu(Bundle.getMessage("MenuOptions"));  // NOI18N
+        JMenu menu = new JMenu(Bundle.getMessage("MenuOptions"));  // NOI18N
         menu.setMnemonic(KeyEvent.VK_O);
+        JMenuBar menuBar = f.getJMenuBar();
         int pos = menuBar.getMenuCount() - 1;  // count the number of menus to insert the TableMenus before 'Window' and 'Help'
         int offset = 1;
         log.debug("setMenuBar number of menu items = {}", pos);  // NOI18N
@@ -227,8 +206,6 @@ public abstract class AbstractLogixNGTableAction<E extends NamedBean> extends Ab
     // Current focus variables
     protected E _curNamedBean = null;
     int conditionalRowNumber = 0;
-
-    protected final Base.PrintTreeSettings _printTreeSettings = new Base.PrintTreeSettings();
 
     // Add E Variables
     JmriJFrame addLogixNGFrame = null;
@@ -723,255 +700,21 @@ public abstract class AbstractLogixNGTableAction<E extends NamedBean> extends Ab
     void browserPressed(String sName) {
         // bean was found, create the window
         _curNamedBean = getManager().getBySystemName(sName);
-        getPrintTreeSettings();
-        makeBrowserWindow(false);
-    }
 
-    void getPrintTreeSettings() {
-        // Set options
-        InstanceManager.getOptionalDefault(UserPreferencesManager.class).ifPresent((prefMgr) -> {
-            _printTreeSettings._printLineNumbers = prefMgr.getSimplePreferenceState(PRINT_LINE_NUMBERS_OPTION);
-            _printTreeSettings._printErrorHandling = prefMgr.getSimplePreferenceState(PRINT_ERROR_HANDLING_OPTION);
-            _printTreeSettings._printNotConnectedSockets = prefMgr.getSimplePreferenceState(PRINT_NOT_CONNECTED_OPTION);
-            _printTreeSettings._printLocalVariables = prefMgr.getSimplePreferenceState(PRINT_LOCAL_VARIABLES_OPTION);
-            _printTreeSettings._printSystemNames = prefMgr.getSimplePreferenceState(PRINT_SYSTEM_NAMES_OPTION);
-            _printTreeSettings._printDisabled = prefMgr.getSimplePreferenceState(PRINT_DISABLED_OPTION);
-            _printTreeSettings._printStartup = true;
-        });
-    }
-
-    /**
-     * Update text in the browser window.
-     * @param browseAll true of browse all LogixNGs, false otherwise
-     */
-    void updateBrowserText(boolean browseAll) {
-        if (_textContent != null) {
-            if (browseAll) {
-                StringWriter writer = new StringWriter();
-                InstanceManager.getDefault(jmri.jmrit.logixng.LogixNG_Manager.class)
-                        .printTree(_printTreeSettings, new PrintWriter(writer), "    ", new MutableInt(0));
-                _textContent.setText(writer.toString());
-            } else {
-                _textContent.setText(getBeanText(_curNamedBean));
-            }
-        }
-    }
-
-    /**
-     * Create and initialize the conditionalNGs browser window.
-     * @param browseAll true of browse all LogixNGs, false otherwise
-     */
-    void makeBrowserWindow(boolean browseAll) {
-        String title = browseAll ? Bundle.getMessage("BrowseAllLogixNGs") : this.getBrowserTitle();
-        JmriJFrame condBrowserFrame = new JmriJFrame(title, false, true);
-
-        condBrowserFrame.addWindowListener(new WindowAdapter() {
-            @Override
-            public void windowClosed(WindowEvent e) {
-                _textContent = null;
-            }
-        });
-
-        Container contentPane = condBrowserFrame.getContentPane();
-        contentPane.setLayout(new BorderLayout());
-
-        // bean header information
-        JPanel topPanel = new JPanel();
-        String tStr;
-        if (browseAll) {
-            tStr = title;
-        } else {
-            tStr = Bundle.getMessage("BrowserLogixNG") + " " + _curNamedBean.getSystemName() + "    " // NOI18N
-                    + _curNamedBean.getUserName() + "    "
-                    + (isEnabled(_curNamedBean)
-                            ? Bundle.getMessage("BrowserEnabled") // NOI18N
-                            : Bundle.getMessage("BrowserDisabled"));  // NOI18N
-        }
-        topPanel.add(new JLabel(tStr));
-        contentPane.add(topPanel, BorderLayout.NORTH);
-
-        // Build the conditionalNGs listing
-        if (browseAll) {
-            StringWriter writer = new StringWriter();
-            InstanceManager.getDefault(jmri.jmrit.logixng.LogixNG_Manager.class)
-                    .printTree(_printTreeSettings, new PrintWriter(writer), "    ", new MutableInt(0));
-            _textContent = new JTextArea(writer.toString());
-        } else {
-            _textContent = new JTextArea(this.getBeanText(_curNamedBean));
-        }
-        if (browseMonoSpace()) {
-            _textContent.setFont(new Font(Font.MONOSPACED, Font.PLAIN, 12));
-        }
-        JScrollPane scrollPane = new JScrollPane(_textContent);
-        contentPane.add(scrollPane);
-
-        JPanel bottomPanel = new JPanel();
-        bottomPanel.setLayout(new BorderLayout());
-        JButton helpBrowse = new JButton(Bundle.getMessage("MenuHelp"));   // NOI18N
-        bottomPanel.add(helpBrowse, BorderLayout.WEST);
-        helpBrowse.addActionListener((ActionEvent e) -> {
-            JmriJOptionPane.showMessageDialog(condBrowserFrame,
-                    Bundle.getMessage("LogixNG_Browse_HelpText"),   // NOI18N
-                    Bundle.getMessage("BrowserHelpTitle"),  // NOI18N
-                    JmriJOptionPane.INFORMATION_MESSAGE);
-        });
-
-        JPanel settingsPanel = getSettingsPanel(browseAll);
-        bottomPanel.add(settingsPanel, BorderLayout.CENTER);
-
-        JButton saveBrowse = new JButton(Bundle.getMessage("BrowserSaveButton"));   // NOI18N
-        saveBrowse.setToolTipText(Bundle.getMessage("BrowserSaveButtonHint"));      // NOI18N
-        bottomPanel.add(saveBrowse, BorderLayout.EAST);
-        saveBrowse.addActionListener((ActionEvent e) -> {
-            saveBrowserPressed();
-        });
-        contentPane.add(bottomPanel, BorderLayout.SOUTH);
-
-        condBrowserFrame.pack();
-        condBrowserFrame.setVisible(true);
-    }  // makeBrowserWindow
-
-    JFileChooser userFileChooser = new jmri.util.swing.JmriJFileChooser(FileUtil.getUserFilesPath());
-
-    /**
-     * Save the bean browser window content to a text file.
-     */
-    void saveBrowserPressed() {
-        userFileChooser.setApproveButtonText(Bundle.getMessage("BrowserSaveDialogApprove"));  // NOI18N
-        userFileChooser.setDialogTitle(Bundle.getMessage("BrowserSaveDialogTitle"));  // NOI18N
-        userFileChooser.rescanCurrentDirectory();
-        // Default to logixNG system name.txt
-        String suggestedFileName = _curNamedBean.getSystemName().replace(':', '_') + ".txt";
-        userFileChooser.setSelectedFile(new File(suggestedFileName));  // NOI18N
-        int retVal = userFileChooser.showSaveDialog(null);
-        if (retVal != JFileChooser.APPROVE_OPTION) {
-            log.debug("Save browser content stopped, no file selected");  // NOI18N
-            return;  // give up if no file selected or cancel pressed
-        }
-        File file = userFileChooser.getSelectedFile();
-        log.debug("Save browser content to '{}'", file);  // NOI18N
-
-        if (file.exists()) {
-            Object[] options = {Bundle.getMessage("BrowserSaveDuplicateReplace"),  // NOI18N
-                    Bundle.getMessage("BrowserSaveDuplicateAppend"),  // NOI18N
-                    Bundle.getMessage("ButtonCancel")};               // NOI18N
-            int selectedOption = JmriJOptionPane.showOptionDialog(null,
-                    Bundle.getMessage("BrowserSaveDuplicatePrompt", file.getName()), // NOI18N
-                    Bundle.getMessage("BrowserSaveDuplicateTitle"),   // NOI18N
-                    JmriJOptionPane.DEFAULT_OPTION,
-                    JmriJOptionPane.WARNING_MESSAGE,
-                    null, options, options[0]);
-            if (selectedOption == 2 || selectedOption == -1) {
-                log.debug("Save browser content stopped, file replace/append cancelled");  // NOI18N
-                return;  // Cancel selected or dialog box closed
-            }
-            if (selectedOption == 0) {
-                FileUtil.delete(file);  // Replace selected
-            }
-        }
-
-        // Create the file content
-        String tStr = Bundle.getMessage("BrowserLogixNG") + " " + _curNamedBean.getSystemName() + "    "  // NOI18N
+        String title = Bundle.getMessage("BrowserLogixNG") + " " + _curNamedBean.getSystemName() + "    " // NOI18N
                 + _curNamedBean.getUserName() + "    "
                 + (isEnabled(_curNamedBean)
-                        ? Bundle.getMessage("BrowserEnabled")    // NOI18N
+                        ? Bundle.getMessage("BrowserEnabled") // NOI18N
                         : Bundle.getMessage("BrowserDisabled"));  // NOI18N
-        try {
-            // Add bean Header inforation first
-            FileUtil.appendTextToFile(file, tStr);
-            FileUtil.appendTextToFile(file, "-".repeat(tStr.length()));
-            FileUtil.appendTextToFile(file, "");
-            FileUtil.appendTextToFile(file, _textContent.getText());
-        } catch (IOException e) {
-            log.error("Unable to write browser content to '{}'", file, e);  // NOI18N
-        }
-    }
 
-    protected JPanel getSettingsPanel(boolean browseAll) {
-        JPanel checkBoxPanel = new JPanel();
-
-        JCheckBox printLineNumbers = new JCheckBox(Bundle.getMessage("LogixNG_Browse_PrintLineNumbers"));
-        printLineNumbers.setSelected(_printTreeSettings._printLineNumbers);
-        printLineNumbers.addChangeListener((event) -> {
-            if (_printTreeSettings._printLineNumbers != printLineNumbers.isSelected()) {
-                _printTreeSettings._printLineNumbers = printLineNumbers.isSelected();
-                InstanceManager.getOptionalDefault(UserPreferencesManager.class).ifPresent((prefMgr) -> {
-                    prefMgr.setSimplePreferenceState(PRINT_LINE_NUMBERS_OPTION, printLineNumbers.isSelected());
+        LogixNGBrowseWindow browseWindow =
+                new LogixNGBrowseWindow(title);
+        browseWindow.getPrintTreeSettings();
+        boolean showSettingsPanel = this instanceof LogixNGTableAction || this instanceof LogixNGModuleTableAction;
+        browseWindow.makeBrowserWindow(browseMonoSpace(), showSettingsPanel, title, _curNamedBean.getSystemName(),
+                (printTreeSettings) -> {
+                        return AbstractLogixNGTableAction.this.getBeanText(_curNamedBean, printTreeSettings);
                 });
-                updateBrowserText(browseAll);
-            }
-        });
-
-        JCheckBox printErrorHandling = new JCheckBox(Bundle.getMessage("LogixNG_Browse_PrintErrorHandling"));
-        printErrorHandling.setSelected(_printTreeSettings._printErrorHandling);
-        printErrorHandling.addChangeListener((event) -> {
-            if (_printTreeSettings._printErrorHandling != printErrorHandling.isSelected()) {
-                _printTreeSettings._printErrorHandling = printErrorHandling.isSelected();
-                InstanceManager.getOptionalDefault(UserPreferencesManager.class).ifPresent((prefMgr) -> {
-                    prefMgr.setSimplePreferenceState(PRINT_ERROR_HANDLING_OPTION, printErrorHandling.isSelected());
-                });
-                updateBrowserText(browseAll);
-            }
-        });
-
-        JCheckBox printNotConnectedSockets = new JCheckBox(Bundle.getMessage("LogixNG_Browse_PrintNotConnectedSocket"));
-        printNotConnectedSockets.setSelected(_printTreeSettings._printNotConnectedSockets);
-        printNotConnectedSockets.addChangeListener((event) -> {
-            if (_printTreeSettings._printNotConnectedSockets != printNotConnectedSockets.isSelected()) {
-                _printTreeSettings._printNotConnectedSockets = printNotConnectedSockets.isSelected();
-                updateBrowserText(browseAll);
-                InstanceManager.getOptionalDefault(jmri.UserPreferencesManager.class).ifPresent((prefMgr) -> {
-                    prefMgr.setSimplePreferenceState(PRINT_NOT_CONNECTED_OPTION, printNotConnectedSockets.isSelected());
-                });
-            }
-        });
-
-        JCheckBox printLocalVariables = new JCheckBox(Bundle.getMessage("LogixNG_Browse_PrintLocalVariables"));
-        printLocalVariables.setSelected(_printTreeSettings._printLocalVariables);
-        printLocalVariables.addChangeListener((event) -> {
-            if (_printTreeSettings._printLocalVariables != printLocalVariables.isSelected()) {
-                _printTreeSettings._printLocalVariables = printLocalVariables.isSelected();
-                updateBrowserText(browseAll);
-                InstanceManager.getOptionalDefault(jmri.UserPreferencesManager.class).ifPresent((prefMgr) -> {
-                    prefMgr.setSimplePreferenceState(PRINT_LOCAL_VARIABLES_OPTION, printLocalVariables.isSelected());
-                });
-            }
-        });
-
-        JCheckBox printSystemNames = new JCheckBox(Bundle.getMessage("LogixNG_Browse_PrintSystemNames"));
-        printSystemNames.setSelected(_printTreeSettings._printSystemNames);
-        printSystemNames.addChangeListener((event) -> {
-            if (_printTreeSettings._printSystemNames != printSystemNames.isSelected()) {
-                _printTreeSettings._printSystemNames = printSystemNames.isSelected();
-                InstanceManager.getOptionalDefault(UserPreferencesManager.class).ifPresent((prefMgr) -> {
-                    prefMgr.setSimplePreferenceState(PRINT_SYSTEM_NAMES_OPTION, printSystemNames.isSelected());
-                });
-                updateBrowserText(browseAll);
-            }
-        });
-
-        JCheckBox printDisabled = new JCheckBox(Bundle.getMessage("LogixNG_Browse_PrintDisabled"));
-        printDisabled.setSelected(_printTreeSettings._printDisabled);
-        printDisabled.addChangeListener((event) -> {
-            if (_printTreeSettings._printDisabled != printDisabled.isSelected()) {
-                _printTreeSettings._printDisabled = printDisabled.isSelected();
-                InstanceManager.getOptionalDefault(UserPreferencesManager.class).ifPresent((prefMgr) -> {
-                    prefMgr.setSimplePreferenceState(PRINT_DISABLED_OPTION, printDisabled.isSelected());
-                });
-                updateBrowserText(browseAll);
-            }
-        });
-
-        if (this instanceof LogixNGTableAction || this instanceof LogixNGModuleTableAction) {
-            checkBoxPanel.add(printLineNumbers);
-            checkBoxPanel.add(printErrorHandling);
-            checkBoxPanel.add(printNotConnectedSockets);
-            checkBoxPanel.add(printLocalVariables);
-            checkBoxPanel.add(printSystemNames);
-            checkBoxPanel.add(printDisabled);
-        }
-
-        return checkBoxPanel;
     }
