+++ java/src/jmri/jmrit/beantable/BeanTableBundle.properties
@@ -799,7 +799,6 @@ The user name must be a valid LogixNG identifier.             \
 </html>
 
 # LogixNG menu
-BrowseAllLogixNGs                    = Browse All LogixNGs
 EnableAllLogixNGs                    = Enable All LogixNGs
 DisableAllLogixNGs                   = Disable All LogixNGs
 # LogixNG Browser
@@ -822,14 +821,6 @@ LogixNG_Table_Browse_Title              = LogixNG Table Browser
 LogixNG_GlobalVar_Browse_Title          = LogixNG Global Variable Browser
 LogixNG_GlobalVar_Browse_Header         = Global Variable: {0} - {1} :: initial type = {2}
 LogixNG_GlobalVar_Browse_Value          = \n    value:\u0020\u0020
-
-LogixNG_Browse_PrintLineNumbers         = Print line numbers
-LogixNG_Browse_PrintErrorHandling       = Print error handling
-LogixNG_Browse_PrintNotConnectedSocket  = Print not connected sockets
-LogixNG_Browse_PrintLocalVariables      = Print local variables
-LogixNG_Browse_PrintSystemNames         = Print system names
-LogixNG_Browse_PrintDisabled            = Print disabled
-LogixNG_Browse_HelpText                 = The text file formatting is based on using a monospace font.
 
 LogixNG_FileButtonHint               = Click to select a file from disk
 LogixNG_CsvFileName                  = CSV file name
