+++ java/src/jmri/jmrit/beantable/BeanTableBundle_cs.properties
@@ -822,13 +822,6 @@ LogixNG_GlobalVar_Browse_Title          = Prohl\u00ed\u017ee\u010d LogixNG Glob\
 LogixNG_GlobalVar_Browse_Header         = Glob\u00e1ln\u00ed prom\u011bnn\u00e1: {0} - {1} :: inicializa\u010dn\u00ed typ = {2}
 LogixNG_GlobalVar_Browse_Value          = \n    hodnota:\u0020\u0020
 
-LogixNG_Browse_PrintLineNumbers         = Tisk \u010d\u00edsel \u0159\u00e1dk\u016f
-LogixNG_Browse_PrintErrorHandling       = Tisk zpracov\u00e1v\u00e1n\u00ed chyb 
-LogixNG_Browse_PrintNotConnectedSocket  = Tisk nep\u0159ipojen\u00fdch soket\u016f
-LogixNG_Browse_PrintLocalVariables      = Tisk lok\u00e1ln\u00edch prom\u011bnn\u00fdch
-LogixNG_Browse_PrintSystemNames         = Tisk syst\u00e9mov\u00fdch n\u00e1zv\u016f
-LogixNG_Browse_HelpText                 = Form\u00e1tov\u00e1n\u00ed textov\u00e9ho souboru je zalo\u017eeno na pou\u017eit\u00ed neproporcion\u00e1ln\u00edho p\u00edsma.
-
 LogixNG_FileButtonHint               = Klikem vyberete soubor z disku
 LogixNG_CsvFileName                  = N\u00e1zev souboru CSV
 LogixNG_ModuleSocketType             = Typ Soketu:
