+++ java/src/jmri/jmrit/logixng/tools/swing/LogixNGMenu.java
@@ -27,6 +27,7 @@ public class LogixNGMenu extends JMenu {
         add(new ImportLogixAction());
         add(new InlineLogixNGsAction());
         add(new WhereUsedAction(Bundle.getMessage("MenuItemWhereUsed")));
+        add(new BrowseAllLogixNGsAction());
     }
 
 //    private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LogixNGMenu.class);
