+++ java/src/jmri/jmrit/logixng/tools/swing/LogixNGSwingBundle.properties
@@ -10,6 +10,7 @@ MenuLogixNGInitializationTable  = LogixNG Initialization table
 MenuImportLogix             = Import Logix
 MenuOpenClipboard           = Open clipboard
 MenuInlineLogixNGs          = Inline LogixNGs
+BrowseAllLogixNGs           = Browse All LogixNGs
 
 ButtonMoveUp                = Move up
 ButtonMoveDown              = Move down
@@ -293,3 +294,11 @@ SaveDuplicateReplace  = Replace
 SaveDuplicateAppend   = Append
 SaveDuplicateTitle    = Duplicate File
 SaveDuplicatePrompt   = File "{0}" already exists, select [{1}] or [{2}]
+
+LogixNG_Browse_PrintLineNumbers         = Print line numbers
+LogixNG_Browse_PrintErrorHandling       = Print error handling
+LogixNG_Browse_PrintNotConnectedSocket  = Print not connected sockets
+LogixNG_Browse_PrintLocalVariables      = Print local variables
+LogixNG_Browse_PrintSystemNames         = Print system names
+LogixNG_Browse_PrintDisabled            = Print disabled
+LogixNG_Browse_HelpText                 = The text file formatting is based on using a monospace font.
