+++ java/src/jmri/jmrit/logixng/tools/swing/LogixNGSwingBundle_cs.properties
@@ -293,3 +293,10 @@ SaveDuplicateReplace  = Nahradit
 SaveDuplicateAppend   = P\u0159ipojit
 SaveDuplicateTitle    = Duplikovat soubor
 SaveDuplicatePrompt   = Soubor "{0}" ji\u017e existuje, vyberte [{1}] nebo [{2}]
+
+LogixNG_Browse_PrintLineNumbers         = Tisk \u010d\u00edsel \u0159\u00e1dk\u016f
+LogixNG_Browse_PrintErrorHandling       = Tisk zpracov\u00e1v\u00e1n\u00ed chyb 
+LogixNG_Browse_PrintNotConnectedSocket  = Tisk nep\u0159ipojen\u00fdch soket\u016f
+LogixNG_Browse_PrintLocalVariables      = Tisk lok\u00e1ln\u00edch prom\u011bnn\u00fdch
+LogixNG_Browse_PrintSystemNames         = Tisk syst\u00e9mov\u00fdch n\u00e1zv\u016f
+LogixNG_Browse_HelpText                 = Form\u00e1tov\u00e1n\u00ed textov\u00e9ho souboru je zalo\u017eeno na pou\u017eit\u00ed neproporcion\u00e1ln\u00edho p\u00edsma.
